Use only routes.js, delete routeObject.js and routes-static.js
