+++ src/index-static.js
@@ -4,7 +4,7 @@ import {Provider} from 'react-redux';
 import createMemoryHistory from 'react-router/lib/createMemoryHistory';
 import RouterContext from 'react-router/lib/RouterContext';
 import match from 'react-router/lib/match';
-import routes from './routes-static';
+import routes from './routes';
 import WithStylesContext from './WithStylesContext';
 import DocumentTitle from 'react-document-title';
 import store from './store';
